+++ framework/db-http/src/test/java/is/codion/framework/db/http/AbstractHttpEntityConnectionTest.java
@@ -377,7 +377,7 @@ abstract class AbstractHttpEntityConnectionTest {
 		Clients.resolveTrustStore();
 		Report.REPORT_PATH.set("report/path");
 		HttpEntityConnection.SECURE.set(true);
-		EntityService.SERALIZATION.set(true);
+		EntityService.SERIALIZATION.set(true);
 		EntityService.KEYSTORE_PATH.set("../../framework/server/src/main/config/keystore.jks");
 		EntityService.KEYSTORE_PASSWORD.set("crappypass");
 
@@ -397,7 +397,7 @@ abstract class AbstractHttpEntityConnectionTest {
 		System.clearProperty(Clients.JAVAX_NET_TRUSTSTORE);
 		System.clearProperty(Clients.JAVAX_NET_TRUSTSTORE_PASSWORD);
 		Report.REPORT_PATH.set(null);
-		EntityService.SERALIZATION.set(false);
+		EntityService.SERIALIZATION.set(false);
 		EntityService.KEYSTORE_PATH.set(null);
 		EntityService.KEYSTORE_PASSWORD.set(null);
 		HttpEntityConnection.SECURE.set(false);