+++ changelog.md
@@ -15,6 +15,8 @@ Codion Change Log
 ### is.codion.framework.domain
 - Column.GetValue and SetParameter moved to common.db as GetValue and SetValue.
 - ColumnDefinition.LEGACY_JDBC moved to Database.
+### is.codion.framework.servlet
+- EntityService.SERIALIZATION config param typo fixed.
 ### is.codion.framework.model
 - AbstractEntityEditor, foreign key based detail editor no longer validates non-present detail entities.
 - AttributeOperands added, AbstractEntityTableModel.EntityFilters now initialize boolean conditions correctly.